refactor: correct path to mocks
